requirements.txt
bidict==0.21.2
click==7.1.2
dnspython==1.16.0
eventlet==0.30.2
Flask==1.1.2
Flask-SocketIO==5.0.1
greenlet==1.0.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pyserial==3.5
python-engineio==4.0.1
python-socketio==5.1.0
six==1.15.0
Werkzeug==1.0.1
